Changed the method used to check for end of tape.
Bug: End of tape (EOD) was identified by failure to read the first data
block for the next file from the tape. However sometimes multiple
file marks can exist next to one another (for untidyly created
backup sets). In these cases there are no data blocks between the
file marks. This causes tapeimgr to think end of tape (EOD) has
been reached when in fact there are more files to be read.

I've modified the method used to determine End of tape. It uses
the data block check first, but if that fails in then does a
secondary check to see if the file number changed after the
data block read. If the file number changed that indicates that
instead of reading a data block the dd fsr command has pushed
the tape to the next file. At end of tape it can't to that (no
more files) so the file number doesn't change.